runtime: fold priority-derivation test into the transaction test module
The dedicated test file was built to model a wire command whose priority
is decoupled from its structure; with priority gone from the Command
trait that concept is unrepresentable, so its PCmd scaffolding
duplicated SeqCommand and its two-identical-runs convergence test had
no failure surface left (real convergence under differing sync orders
is covered by the DSL suites).

Keep the one check with a real regression surface: add_command accepts
any Priority, so a plumbing bug (wrong variable in add_single, an
indexing bug in Segment::get_priority) would type-check and even
converge, since every peer would derive the same wrong value. Walk a
committed graph exercising all four priority kinds and assert each
stored priority equals an independent derivation.
